fastapi-jwt-auth-mongodb/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
motor==3.6.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.18
starlette==0.40.0
